Update apg-go commands in README
This commit replaces all instances of apg with apg-go in the examples of the README.md file, making them consistent and accurate. Additionally, a new feature for mobile-friendly character grouping has been hooked in. This feature improves user experience, especially for mobile users, by organising the generated passwords in a "keypad-order" grouping.