Add Go install instructions to README
This commit includes detailed Go install instructions for the apg-go project into the README.md. The instructions verify Go version, describe the installation of the latest version and retrieval of a specific version, and also details on how to execute apg-go. This additional information allows users with Go installed in their systems to easily use and test the apg-go project.